minor bugfixes as I test.  moved some utility functions to utils.py
